fix: only memory address apply the uint64_t, the indices of memories still are uint32_t

+++ lib/executor/engine/proxy.cpp
@@ -710,7 +710,7 @@ Expect<void> Executor::proxyTableFill(Runtime::StackManager &StackMgr,
 }
 
 Expect<uint64_t> Executor::proxyMemGrow(Runtime::StackManager &StackMgr,
-                                        const uint64_t MemIdx,
+                                        const uint32_t MemIdx,
                                         const uint64_t NewSize) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
   assuming(MemInst);
@@ -729,15 +729,15 @@ Expect<uint64_t> Executor::proxyMemGrow(Runtime::StackManager &StackMgr,
 }
 
 Expect<uint64_t> Executor::proxyMemSize(Runtime::StackManager &StackMgr,
-                                        const uint64_t MemIdx) noexcept {
+                                        const uint32_t MemIdx) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
   assuming(MemInst);
   return MemInst->getPageSize();
 }
 
 Expect<void>
-Executor::proxyMemInit(Runtime::StackManager &StackMgr, const uint64_t MemIdx,
-                       const uint64_t DataIdx, const uint64_t DstOff,
+Executor::proxyMemInit(Runtime::StackManager &StackMgr, const uint32_t MemIdx,
+                       const uint32_t DataIdx, const uint64_t DstOff,
                        const uint64_t SrcOff, const uint64_t Len) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
   assuming(MemInst);
@@ -747,7 +747,7 @@ Executor::proxyMemInit(Runtime::StackManager &StackMgr, const uint64_t MemIdx,
 }
 
 Expect<void> Executor::proxyDataDrop(Runtime::StackManager &StackMgr,
-                                     const uint64_t DataIdx) noexcept {
+                                     const uint32_t DataIdx) noexcept {
   auto *DataInst = getDataInstByIdx(StackMgr, DataIdx);
   assuming(DataInst);
   DataInst->clear();
@@ -755,8 +755,8 @@ Expect<void> Executor::proxyDataDrop(Runtime::StackManager &StackMgr,
 }
 
 Expect<void> Executor::proxyMemCopy(Runtime::StackManager &StackMgr,
-                                    const uint64_t DstMemIdx,
-                                    const uint64_t SrcMemIdx,
+                                    const uint32_t DstMemIdx,
+                                    const uint32_t SrcMemIdx,
                                     const uint64_t DstOff,
                                     const uint64_t SrcOff,
                                     const uint64_t Len) noexcept {
@@ -770,7 +770,7 @@ Expect<void> Executor::proxyMemCopy(Runtime::StackManager &StackMgr,
 }
 
 Expect<void> Executor::proxyMemFill(Runtime::StackManager &StackMgr,
-                                    const uint64_t MemIdx, const uint64_t Off,
+                                    const uint32_t MemIdx, const uint64_t Off,
                                     const uint8_t Val,
                                     const uint64_t Len) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
@@ -779,7 +779,7 @@ Expect<void> Executor::proxyMemFill(Runtime::StackManager &StackMgr,
 }
 
 Expect<uint64_t> Executor::proxyMemAtomicNotify(Runtime::StackManager &StackMgr,
-                                                const uint64_t MemIdx,
+                                                const uint32_t MemIdx,
                                                 const uint64_t Offset,
                                                 const uint64_t Count) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
@@ -789,7 +789,7 @@ Expect<uint64_t> Executor::proxyMemAtomicNotify(Runtime::StackManager &StackMgr,
 
 Expect<uint64_t>
 Executor::proxyMemAtomicWait(Runtime::StackManager &StackMgr,
-                             const uint64_t MemIdx, const uint64_t Offset,
+                             const uint32_t MemIdx, const uint64_t Offset,
                              const uint64_t Expected, const int64_t Timeout,
                              const uint32_t BitWidth) noexcept {
   auto *MemInst = getMemInstByIdx(StackMgr, MemIdx);
